allow selecting colour when adding a spirit
Before this commit, the host has no choice, so colours are randomly
selected.

If specific colours are desired, then they have to ask an admin to
change the colours after the spirit is added.

Wouldn't it be nice to be able to specify the colour when adding the
spirit and not have to bother an admin?