implement Dark Fire Shadows' permanent moon with presence
Before this change, it's possible for the player to accidentally
decrement their permanent moon and lose it.

With this change, the advantage is that this feels more in line with
other bonus elements granted by the track: The player is always entitled
to them and cannot accidentally lose them due to misclick.

No migration is needed for existing games with a Dark Fire Shadows
because both init_permanent_elements and add_player are only run on
player creation. Existing instances of Dark Fire Shadows continue to
use permanent_elements and don't get this new invisible presence. Only
new instances of Dark Fire Shadows get this.