choose_card: streamline Boon of Reimagining
By not sending the cards to the discard if there are 5 cards remaining
in the selection, players can just click gain on the two cards they
want.

Before this code change, gaining their first choice will send the
selection to the discard, and then the player must gain the second out
of the discard, This works but is a poor experience.