fail loudly if any card has duplicate elements
This avoids data-entry errors, and provides future-proofing.

Currently the code makes an unenforced assumption about released
material (namely that no card has duplicate elements).

If we do not enforce the assumption, someone attempting to add a future
card that has two of any element will observe the site silently do the
wrong thing, or worse, they will not notice at all.

By enforcing the assumption and erroring if it's violated, we
future-proof the site. Now the site will error, and by searching for the
error message in the codebase, they'll be led to the accompanying
function and comment that explains what needs to be changed. They'll
make the change and their new card will work as they expect.

closes https://github.com/nathanj/spirit-island-pbp/issues/39