allow players to undo gaining cards
Sometimes players accidentally hit the button and then need an admin's
help to reshuffle the cards back in the deck.

This change will let them do it themselves.

It is assumed that players will only use this when they genuinely made a
mistake rather than to cheat on their card gains.

This doesn't help if someone accidentally hit the Take 1 button though,
since then the card goes right to their hand instead of to the
selection. Something else will be needed for that case.