Integrate seeddb into the database migration flow
Closes https://github.com/nathanj/spirit-island-pbp/issues/90

When creating a brand new enlistment and a new DB for the first time,
the seeddb command was failing due to the addition of a field to the
Card model in later migrations causing it to attempt to use the most
recent Card model to seed the DB when the state of the DB at the time
only contained the initial Card definitions.

This change merges the initial population of the database into the
initial database migration itself, removing the need for the seeddb
command entirely and also helping future-proof the database since it
means that we will no longer have to worry about keeping future
migrations as compatible with the initial models.