remove vertical gap for narrower screen sizes
Before this commit, when the spirit panel is resized for narrower screen
sizes, the height of the vertical bounding box is not changed. This is
just how CSS scale works. This causes a vertical gap to appear between
the spirit panel and the ready button beneath it. The vertical gap
doesn't harm functionality but takes up unnecessary space.

This commit removes the vertical gap by setting the height of the
containing div according to the scaling factor.
Note the necessity of supporting a double-height spirit panel as well,
for Covets Gleaming Shards of Earth's hoard panel.
A small margin is added below so that the ready button isn't directly
touching the spirit panel.

Closes https://github.com/nathanj/spirit-island-pbp/issues/49