reimplement exploratory Bringer of Dreams and Nightmares as aspect
This is safe to do because exploratory is not intended to combine with
any other aspect.

This indirectly fixes the problem where exploratory Bringer doesn't
start with any uniques.
An alternative solution was to just use the add cards function that's
normally used for aspects, but that leaves in a little bit of duplicate
code/data for exploratory Bringer, which is unnecessary since so much is
the same.

close https://github.com/nathanj/spirit-island-pbp/issues/54