add scroll workaround (change all buttons to a)
This sets a cookie that is used to choose whether to use `<button>` or
`<a>`

Using `<a>` has some undesirable behaviour (the browser attempts to
highlight the text when clicking it), but so far this is the only
reliable way that's been found to stop the UI from scrolling improperly
on some browsers/platforms (Chrome-based browsers on Windows). So, we'll
allow individual players to choose what works best for them.

https://github.com/nathanj/spirit-island-pbp/issues/131