simplify host draw to one button + select
Previously it had unclear behaviour on hitting enter, as there are two
buttons. So which card do you expect it to discard?

We'll simplify this by using a select instead. This is acceptable
because major powers are only drawn for Shattered Fragments of Power, so
it's okay for it to be an extra input to access.

Closes https://github.com/nathanj/spirit-island-pbp/issues/143