pyproject.toml: replace requests-async with just requests
This project has no instances of `import requests_async` and thus was
not using requests-async. It has never had any instances of this import
line in its entire history.

However, it does have one instance of `import requests` in bot.py, so we
need to put requests in requests-async's place.

Closes https://github.com/nathanj/spirit-island-pbp/issues/87