bot: don't crash if we couldn't load the emojis of the right guild
This is helpful for new site administrators setting up their new
instance of the bot.

New site admins who hadn't invited the bot to the guild yet were getting
"'NoneType' object is not iterable", but emojis aren't critical to the
functionality of the bot, as evidenced in the load_emojis function where
it simply warns, rather than crashing, if it couldn't find the emoji for
a few particular spirits. Better to print a warning.