move base template style to css file
This allows browsers to cache this file which doesn't change, and means
we no longer have to send this content with every single request.

However, this change is pretty low impact. A full page load is expected
to be a few megabytes, and this is about a 1 kilobyte savings, so it's
less than 0.1%.