update htmx to 2.0.6
main motivation is to benefit from the new default of true for
htmx.config.selfRequestsOnly. This setting was added in 1.9.5 so even if
we don't update to 2.0.6 we would still want to update to 1.9.5 for
this. We want this setting because in no case do we ever want to include
HTML from an external site, so we want htmx to enforce this for us.

Disadvantages are:

* larger file size (relying on browsers to cache the file to mitigate
  this disadvantage)
* dropped support of Internet Explorer (relying on our players to be
  reasonable and not using outdated browser)

migration guide indicates nothing we're using needs to be updated
https://htmx.org/migration-guide-htmx-1/

https://github.com/bigskysoftware/htmx/blob/master/CHANGELOG.md

The sha384sum of this file is:
024a9fadb8ff1e9355a3c935d525c16fa4e50569975e5610ac24aa1169b22897be8439b767f0765951b8b26c01911566

this matches the integrity value given on https://htmx.org/docs/ which
is is in base64:
Akqfrbj/HpNVo8k11SXBb6TlBWmXXlYQrCSqEWmyKJe+hDm3Z/B2WVG4smwBkRVm

```
$ echo Akqfrbj/HpNVo8k11SXBb6TlBWmXXlYQrCSqEWmyKJe+hDm3Z/B2WVG4smwBkRVm | base64 -d | xxd -p -c96
024a9fadb8ff1e9355a3c935d525c16fa4e50569975e5610ac24aa1169b22897be8439b767f0765951b8b26c01911566
```