discard top N powers: use radio instead of select
Not only does this save one click for selecting majors, but it also
works around the problem where clicking buttons that reload player.html
on Chrome would cause the page to jump to the bottom. This behaviour
only happened if there was a `<select>` anywhere on the pgae, and this
is the last one remaining. Removing it works around it, though we don't
at this time know why it's happening.

Closes https://github.com/nathanj/spirit-island-pbp/issues/131