admin: Use filter_horizontal on minor_deck, major_deck, discard_pile
Found this setting in
https://docs.djangoproject.com/en/5.2/ref/contrib/admin/

By default the many-to-many uses a `<select multiple>` which has been
very cumbersome, since the most common operation is to add/remove one
item, and you need to have the right key held down to do this.

With the filter_horizontal, cards in the deck are on the right, cards
not in the deck are on the left, and you can filter, select, and confirm
which cards should be added/removed.