migrations: Add blank=True to Card's spirit
Should have been done with c5d4db76e63bc3865c995f54e0b219d30fa34f40

This has no effect on the database, so the author mistakenly thought
there was no need to have a migration.

The blank property only relates to form validation, such as in the admin
interface. This allows the spirit to be not specified in such a form.