support CORS (Cross-Origin Resource Sharing)
The site administrator will be able to allow external sites to make
requests to this site.

To test, send the following request to the server:

    curl -H "Access-Control-Request-Method: GET" -H "Origin: https://www.example.com" --head http://localhost:8000 | grep -i 'access-control-allow-origin\|$'

Look for the presence or absence of access-control-allow-origin header
in the response.

The header will only be present if the CORS_ALLOWED_ORIGINS was set to
include https://www.example.com

Closes https://github.com/nathanj/spirit-island-pbp/issues/162 assuming
that the site admininstrators also allows the external site there
mentioned.