Game: Add ability to always suffix filenames
In 225228f00292dafd1ac05247136c6b3de22c2d3e we install django-cleanup to
clean up old screenshots. However for hosts who reuse the same filenames
we need to deal with browser caching when the files eventually get
stored in the same location.

This adds a setting that can be manually toggled to always suffix
filenames, avoiding this issue.