remove dependency load-dotenv
This dependency does absolutely nothing for this project.

It only operates if LOAD_DOTENV is set to true, yes, or 1, but nobody
using this project has been instructed to set this variable, nor do we
set in any production instance of this project.

Even if it were to operate, it would call `load_dotenv`, which is
redundant because both the bot and site already call this function.