GamePlayer/elements: make elements a cached_property
elements queries in-play cards and presences each time it's called, so
we want to cache its result.

Before this commit, a player view makes 4+N queries for in-play cards and
5+N queries for presences (N is the number of cards in hand, play,
selection, and impending).

After this commit, a player view now makes three queries for in-play
cards and four queries for presences, not increasing based on how many
cards they have.

Elements is called for these purposes, and all but one of them can now
use the cache:

* showing the temporary elements UI
* computing spirit thresholds
* computing thresholds of power cards, once per card in hand, play,
  selection, and impending

Thus, for a newly-created spirit with four cards, 5/6 calls to elements
now use the cache.
The savings only go up the more cards a spirit has in hand.