Game: provide player count and player summary properties
We only need the full info of the player whose tab we are actively
showing. For all other players, we only need just enough info to display
their tab.

Queries before this commit:

* one to get the ID of the first player to show their tab
* three to count the number of players:
  * one to check if it's non-zero to decide whether to show tabs
  * two for the discard N cards from deck button
* one to show the full info of the current tab's player
* one (full info, unnecessary) to show the tab list

Queries after this commit:

* one to get the ID of the first player to show their tab
* one to count the number of players (populates player_count)
* one to show the full info of the current tab's player
* one (summary) to show the tab list

Two duplicate queries counting the number of players have thus been
removed entirely, and one that was previously getting full info now only
gets the info it needs.

You'd think the count and summary queries could be combined, but note
that the tabs are a different HTTP request from the main game page and
therefore can't share the same instance of Game.

Whether this is possible to resolve is for future investigation; it's
outside the scope of this commit.