Game: provide player summary property
We only need the full info of the player whose spirit panel we are
actively showing. For all other players, we only need just enough info
to display their tab.

Before this commit we were fetching the full info of all players and
constructing model objects.
After this commit we only fetch the requested fields, and construct
named tuples. The interface provided to the template is essentially the
same, and we've provided all the info it expect to receive, so it
continues to work.