try_match_spirit: Only query for the ID, not the entire model
`first().id` loads the entire model, which is unnecessary because we
only need the ID.

https://docs.djangoproject.com/en/5.2/ref/models/querysets/ explains
that `first()` is one of the methods that doesn't return a `QuerySet`.

Instead, we should use `values_list` which creates a `QuerySet` that
will return just the specified fields when evaluated.

This allows us to remove the unnecessary `exists()` query as well, since
`first()` will return `None` if there are no matches, which is exactly
the semantics we want, because that will cause us to try to use the next
search criterion. Previously `first().id` would crash because of the
`.id` lookup on a `None`.