view_game: Query only the ID of the first player, not the entire model
`first().id` loads the entire model, which is unnecessary because we
only need the ID.

https://docs.djangoproject.com/en/5.2/ref/models/querysets/ explains
that `first()` is one of the methods that doesn't return a `QuerySet`.

Instead, we should use `values_list` which creates a `QuerySet` that
will return just the specified fields when evaluated.

This allows us to remove the unnecessary `exists()` query as well, since
`first()` will return `None` if there are no players, which is fine as
we just need it not to crash. Previously `first().id` would crash
because of the `.id` lookup on a `None`.

Before this commit, these two queries:

* one checking for the existence of any players
* one select of all fields of the first player

Turn into a single query:

* one select of only the ID of the first player