drop the separate spirit Exploratory Bringer (it's an aspect now)
There was no way to access this spirit other than manually setting it in
the admin UI, as we do not include it in the list of spirits in game
setup. Game setup gives you access to the aspect, which was added in
53082b06c266cc5677df13f0d037f48a6d3531fb.

We could have done this when we switched to the aspect but didn't at the
time. Let's do it now that it's been a few months. The game setup screen
keeps posting a log message

    Warning: missing spirits {'Exploratory Bringer'}

So let's get rid of that error, as well as make clear that the supported
way to use exploratory Bringer is the aspect.