don't set a selection if the player already has one active
Otherwise, cards in the previous selection will disappear from the game
and no longer be accessible.

We already enforced this in the UI, but now this adds the same check
server-side.

A player might accidentally do this if they have multiple tabs open to
the same game, or in the case of a slow network.