GamePlayer: remove base_energy_per_turn field
Consider the two reasons why this might make sense as a field of
GamePlayer:

1. Can each GamePlayer have a separate base energy per turn, changing
   independently of the spirit they're playing? In this case, no, the
   rules of the game don't allow this. The base energy gain is always
   determined by the spirit they play.
2. Do we want to speed up read queries, such that we don't need to know
   what spirit a player is playing to know the base energy per turn? In
   database design this is called denormalisation, a choice to store
   redundant data to speed up reads.

Let's examine the second question.

There is one time we use base energy per turn, namely when calculating
energy gain. There are three times we calculate energy gain:

* When creating a log message. In this case, we already look up the
  spirit to put its name in the log message.
* When displaying the spirit's tab. In this case, we already look up the
  spirit's name to determine what spirit panel to show.
* When gaining energy for the turn. It's true that this does add one
  lookup, but the spirit table has a small number of rows and columns
  and the gain energy operation doesn't make many queries compared to
  others (7 selects 1 update after this change, vs showing player tab 14
  selects)

So most of the time, denormalisation doesn't get us anything. It's only
saving a query in the third case (gaining energy for the turn), and we
can make this better by not looking up the spirit if there is any energy
presence uncovered.

Since denormalisation is only giving very small benefits, it seems
better to avoid storing the redundant data.