admin: only change cards in debug
Cards are shared across all games. If anyone edits one, the change
applies to all games. The admin interface, with no controls, is an
inappropriate place to edit cards in production. To prevent corrupt
data, we cannot allow it.

On the other hand, if someone really wants to corrupt their own data on
their private debug instance, that's their own decision and they can
deal with the consequences.