bot: deprecate the pin and unpin (reply) commands
The admins of the main community using the bot have informed us that
they've allowed hosts to pin messages under their own power, so the bot
is no longer needed to serve this function.

However, unpin N is still useful as a convenience.

For an unspecified transitionary period, let's choose to still provide
the pin functionality, but educate hosts that they can do so themselves.